Rename project "PBI-PBIP-filters" -> "PBIP-tools"
